Changes to scripts
